backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
python-dotenv==1.0.1
lunar-python==1.3.10
pydantic==2.6.0
pydantic-settings==2.1.0
cryptography
httpx==0.26.0
langchain==1.2.6
langchain-openai==1.1.7
langchain-community==0.4.1
chromadb==1.4.1
sentence-transformers==5.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7